backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Amazing Market Backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "mysql-connector-python>=9.6.0",
    "python-dotenv>=1.2.2",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "coverage>=7.0.0",
    "httpx>=0.27.0",
]

[tool.coverage.run]
branch = true
source = ["app"]
omit = ["*/tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
